Moving logger to util directory
