+++ src/label/creator.py
@@ -1,7 +1,7 @@
 import qrcode
 from datetime import datetime, timedelta
 from time import time
-from logger import get_logger
+from src.util.logger import get_logger
 import json
 import math
 from PIL import Image, ImageDraw, ImageFont
